перенес channel в активити
