+++ app/build.gradle
@@ -49,9 +49,11 @@ android {
 
 dependencies {
 
+    implementation 'com.google.firebase:firebase-messaging:22.0.0'
+    implementation 'com.google.firebase:firebase-core:19.0.1'
+
     //Room
     implementation 'androidx.room:room-runtime:2.3.0'
-    implementation 'com.google.firebase:firebase-messaging-ktx:22.0.0'
     kapt 'androidx.room:room-compiler:2.3.0'
 
     //Picasso
