+++ app/src/main/java/ru/geekbrains/androidwithkotlin/view/MainActivity.kt
@@ -1,10 +1,19 @@
 package ru.geekbrains.androidwithkotlin.view
 
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.content.Context
+import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import ru.geekbrains.androidwithkotlin.R
+import ru.geekbrains.androidwithkotlin.cloudmessage.CHANNEL_ID
 import ru.geekbrains.androidwithkotlin.databinding.MainActivityBinding
 
 class MainActivity : AppCompatActivity() {
@@ -20,6 +29,35 @@ class MainActivity : AppCompatActivity() {
                 .replace(binding.container.id, MainFragment.newInstance())
                 .commitNow()
         }
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O){
+            val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+            createNotificationChannel(notificationManager)
+        }
+
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Log.w("FIREBASEMSG", "Fetching FCM registration token failed", task.exception)
+                return@OnCompleteListener
+            }
+
+            // Get new FCM registration token
+            val token = task.result
+
+            Log.d("FIREBASEMSG", token!!)
+        })
+
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    private fun createNotificationChannel(notificationManager: NotificationManager) {
+        val channelName = "Channel name"
+        val descriptionText = "Channel description"
+        val importance = NotificationManager.IMPORTANCE_DEFAULT
+        val channel = NotificationChannel(CHANNEL_ID, channelName, importance).apply {
+            description = descriptionText
+        }
+        notificationManager.createNotificationChannel(channel)
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
